+++ lib/vcr/middleware/rack.rb
@@ -4,8 +4,10 @@ module VCR
       include Common
 
       def call(env)
-        VCR.use_cassette(*cassette_arguments(env)) do
-          @app.call(env)
+        Thread.exclusive do
+          VCR.use_cassette(*cassette_arguments(env)) do
+            @app.call(env)
+          end
         end
       end
     end