+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `:ignore_cassettes` option to `VCR.turn_off!`.  This causes
   cassette insertions to be ignored rather than to trigger an error.
   Patch provided by [Justin Smestad](https://github.com/jsmestad).
+* Fix rack middleware to make it threadsafe.
 
 ## 1.9.0 (April 14, 2011)